Finished running split_kmer_counts for alternative haplotype kmers

+++ src/collect_kmers/organize_directories.sh
@@ -12,15 +12,17 @@
 
 cd /home/groups/dpwall/briannac/alt_haplotypes/intermediate_files/kmers
 while read SAMPLE _; do
-    echo "running" $SAMPLE
-    mkdir $SAMPLE
-    mv $SAMPLE.query_counts.unmapped_reads.txt.gz $SAMPLE/$SAMPLE.query_counts.unmapped_reads.txt.gz
-    mv $SAMPLE.query_counts.unmapped_reads.txt $SAMPLE/$SAMPLE.query_counts.unmapped_reads.txt
-    mv $SAMPLE.jellyfish.unmapped_reads.jf $SAMPLE/$SAMPLE.jellyfish.unmapped_reads.jf
-    mv $SAMPLE.jellyfish.unmapped_reads.fa $SAMPLE/$SAMPLE.jellyfish.unmapped_reads.fa
-    mv $SAMPLE.query_counts.unmapped_reads.done $SAMPLE/$SAMPLE.query_counts.unmapped_reads.done
-    for i in {0..2644}; do
-        N=$(printf "%04g" $i)
-        mv $SAMPLE.query_counts.unmapped_reads.$N.txt $SAMPLE/$SAMPLE.query_counts.unmapped_reads.$N.txt
-    done
+    if [ -f "$SAMPLE.query_counts.unmapped_reads.2644.txt" ]; then
+        echo "running" $SAMPLE
+        mkdir $SAMPLE
+        mv $SAMPLE.query_counts.unmapped_reads.txt.gz $SAMPLE/$SAMPLE.query_counts.unmapped_reads.txt.gz
+        mv $SAMPLE.query_counts.unmapped_reads.txt $SAMPLE/$SAMPLE.query_counts.unmapped_reads.txt
+        mv $SAMPLE.jellyfish.unmapped_reads.jf $SAMPLE/$SAMPLE.jellyfish.unmapped_reads.jf
+        mv $SAMPLE.jellyfish.unmapped_reads.fa $SAMPLE/$SAMPLE.jellyfish.unmapped_reads.fa
+        mv $SAMPLE.query_counts.unmapped_reads.done $SAMPLE/$SAMPLE.query_counts.unmapped_reads.done
+        for i in {0..2644}; do
+            N=$(printf "%04g" $i)
+            mv $SAMPLE.query_counts.unmapped_reads.$N.txt $SAMPLE/$SAMPLE.query_counts.unmapped_reads.$N.txt
+        done
+    fi
 done < $MY_HOME/general_data/samples_and_batches.tsv
